+++ edb/server/protocol/auth_ext/data.py
@@ -47,7 +47,7 @@ class UserInfo:
     phone_number: Optional[str] = None
     phone_number_verified: Optional[bool] = None
     address: Optional[dict[str, str]] = None
-    updated_at: Optional[int] = None
+    updated_at: Optional[float] = None
 
     def __str__(self) -> str:
         return self.sub
